Check if an object is a character

Check if an object is a character

lama_check_character(obj, len_1 = TRUE, allow_null = FALSE,
  allow_empty = FALSE, allow_na = FALSE,
  err_handler = composerr("Error in 'lama_check_character"))

Arguments

obj
An object that should be checked if it has the right length

len_1
A flag that tells if obj should be of length 1

allow_null
A flag that tells if NULL is allowed.

allow_empty
A flag that tells if "" is allowed.

allow_na
A flag that tells if NA is allowed.

err_handler
An error handling function, which should be used