Merge multiple label lexicas into one

This function takes multiple lama_dictionary class objects and merges them together into a single lama_dictionary class object. In case some class objects have entries with the same name, the class objects passed in later overwrite the class objects passed in first (e.g. in lama_merge(x, y, z): The lexicon z overwrites x and y. The lexicon y overwrites x).

lama_merge(..., show_warnings = TRUE)

# S3 method for lama_dictionary
lama_merge(..., show_warnings = TRUE)

Arguments

...
Two or more lama_dictionary class objects, which should be merged together.

show_warnings
A logical flag that defines, whether warnings should be shown (TRUE) or not (FALSE).

Value

The merged lama_dictionary class object

See also

Examples

  # initialize lama_dictinoary
  dict_1 <- new_lama_dictionary(
    subject = c(en = "English", ma = "Mathematics"),
    result = c("1" = "Very good", "2" = "Good", "3" = "Not so good")
  )
  dict_2 <- new_lama_dictionary(
    result = c("1" = "Super", "2" = "Fantastic", "3" = "Brilliant"),
    grade = c(a = "Primary School", b = "Secondary School")
  )
  dict_3 <- new_lama_dictionary(
    country = c(en = "England", "at" = "Austria", NA_ = "Some other country")
  )
  dict <- lama_merge(dict_1, dict_2, dict_3)

#> Warning: The following lama_dictionary entries will be overwritten: 'result'.

  # The lama_dictionary now contains the translations
  # 'subject', 'result', 'grade' and 'country'
  # The translation 'result' from 'dict_1' was overwritten by the 'result' in 'dict_2'
  dict

#> 
#> --- lama_dictionary ---
#> Variable 'subject':
#>            en            ma 
#>     "English" "Mathematics" 
#> 
#> Variable 'result':
#>           1           2           3 
#>     "Super" "Fantastic" "Brilliant" 
#> 
#> Variable 'grade':
#>                  a                  b 
#>   "Primary School" "Secondary School" 
#> 
#> Variable 'country':
#>                   en                   at                  NA_ 
#>            "England"            "Austria" "Some other country" 
#>