Change or append a variable translation to an existing lama_dictionary object

The functions lama_mutate() and lama_mutate_() alter a lama_dictionary object. They either alter or append a translation to a lama_dictionary object. The function lama_mutate() uses named arguments to assign the translations to the new names (similar to dplyr::mutate()), whereas the function lama_mutate_() is takes a character string key holding the name to which the translation should be assigned and a named character vector translation holding the actual translation mapping.

lama_mutate(.data, ...)

# S3 method for lama_dictionary
lama_mutate(.data, ...)

lama_mutate_(.data, key, translation)

# S3 method for lama_dictionary
lama_mutate_(.data, key, translation)

Arguments

.data
A lama_dictionary object

...
One or more unquoted expressions separated by commas. Use named arguments, e.g. new_transation_name = c(a = "A", b = "B"), to set translations (named character vectors) to new translation names. It is also possible use complex expressions as long as the resulting object is a valid translation object (named character vector). Furthermore, it is possible to use translation names that are already existing in the dictionary, in order to modify them (e.g. new_translation = c(v = "V", w = "W", old_translation, z = "Z"), where old_translation = c(x = "X", y = "Y")).

key
The name of the variable translation that should be altered. It can also be variable translation name that does not exist yet.

translation
A named character vector holding the new variable translation that should be assigned to the name given in argument key. The names of the character vector translation correspond to the original variable values that should be replaced by the new labels. The values in the character vector translations are the labels that should be assigned to the original values.

Value

An updated lama_dictionary class object.

See also