Assign new labels to a variable of a data.frame

The functions lama_translate() and lama_translate_() take a factor, a vector or a data.frame and convert one or more of its categorical variables (not necessarely a factor variable) into factor variables with new labels. The function lama_translate() uses non-standard evaluation, whereas lama_translate_() is the standard evaluation alternative. The functions lama_translate() and lama_translate_() require different arguments, depending on the data type passed into argument .data. If .data is of type character, logical, numeric or factor, then the arguments col and col_new are omitted, since those are only necessary in the case of data frames.

lama_translate(.data, dictionary, ..., keep_order = FALSE)

# S3 method for data.frame
lama_translate(.data, dictionary, ...,
  keep_order = FALSE)

# S3 method for default
lama_translate(.data, dictionary, ...,
  keep_order = FALSE)

lama_translate_(.data, dictionary, translation, col = translation,
  col_new = col, keep_order = FALSE, ...)

# S3 method for data.frame
lama_translate_(.data, dictionary, translation,
  col = translation, col_new = col, keep_order = FALSE, ...)

# S3 method for default
lama_translate_(.data, dictionary, translation, ...,
  keep_order = FALSE)

Arguments

.data
Either a data frame, a factor or an atomic vector.

dictionary
A lama_dictionary object, holding the translations for various variables.

...
Only used by lama_translate(). Each argument in ... is an unquoted expression and defines a translation. Use unquoted arguments that tell which translation should be applied to which column and which column name the relabelled variable should be assigned to. E.g. lama_translate(.data, dict, Y1 = TRANS1(X1), Y2 = TRANS2(Y2)) to apply the translations TRANS1 and TRANS2 to the data.frame columns X1 and X2 and save the new labelled variables under the column names Y1 and Y2. There are also two abbreviation mechanisms available: The argument assignement FOO(X) is the same as X = FOO(X) and FOO is an abbreviation for FOO = FOO(FOO). In case, .data is not a data frame but a plain factor or an atomic vector, then the argument ... must be a single unquoted translation name (e.g. lama_translate(x, dict, TRANS1), where x is a factor or an atomic vector and TRANS1 is the name of the translation, which should be used to assign the labels to the values of x.)

keep_order
A boolean vector of length one or the same length as the number of translations. If the vector has length one, then the same configuration is applied to all variable translations. If the vector has the same length as the number of arguments in ..., then the to each variable translation there is a corresponding boolean configuration. If a translated variable in the data.frame is a factor variable, and the corresponding boolean configuration is set to TRUE, then the the order of the original factor variable will be preserved.

translation
A character vector holding the names of the variable translations which should be used for assigning new labels to the variable. This names must be a subset of the translation names returned by names(dictionary).

col
Only used if .data is a data frame. The argument col must be a character vector of the same length as translation holding the names of the data.frame columns that should be relabelled. If omitted, then it will be assumed that the column names are the same as the given translation names in the argument translation.

col_new
Only used if .data is a data frame. The argument col must be a character vector of the same length as translation holding the names under which the relabelled variables should be stored in the data.frame. If omitted, then it will be assumed that the new column names are the same as the column names of the original variables.

Value

An extended data.frame, that has a factor variable holding the assigned labels.

Examples

  # --- initialize lama_dictinoary ---
  dict <- new_lama_dictionary(
    subject = c(en = "English", ma = "Mathematics"),
    result = c("1" = "Very good", "2" = "Good", "3" = "Not so good")
  )
  df <- data.frame(
    pupil = c(1, 1, 2, 2, 3),
    subject = c("en", "ma", "ma", "en", "en"),
    res = c(1, 2, 3, 2, 2)
  )
  # --- label data.frame variables ---
  # Example-1 (data.frame): Full length assignment
  # (apply translation 'subject' to column 'subject' and save it to column 'subject_new')
  # (apply translation 'result' to column 'res' and save it to column 'res_new')
  df_new <- lama_translate(
    df,
    dict,
    sub_new = subject(subject),
    res_new = result(res)
  )
  str(df_new)

#> 'data.frame':	5 obs. of  5 variables:
#>  $ pupil  : num  1 1 2 2 3
#>  $ subject: Factor w/ 2 levels "en","ma": 1 2 2 1 1
#>  $ res    : num  1 2 3 2 2
#>  $ sub_new: Factor w/ 2 levels "English","Mathematics": 1 2 2 1 1
#>  $ res_new: Factor w/ 3 levels "Very good","Good",..: 1 2 3 2 2

  # Example-2 (data.frame): Overwrite the original columns (abbreviation)
  # (apply translation 'subject' to column 'subject' and save it to column 'subject')
  # (apply translation 'result' to column 'res' and save it to column 'res')
  df_new_overwritten <- lama_translate(
    df,
    dict,
    subject(subject),
    result(res)
  )
  str(df_new_overwritten)

#> 'data.frame':	5 obs. of  3 variables:
#>  $ pupil  : num  1 1 2 2 3
#>  $ subject: Factor w/ 2 levels "English","Mathematics": 1 2 2 1 1
#>  $ res    : Factor w/ 3 levels "Very good","Good",..: 1 2 3 2 2

  # Example-3 (data.frame): Abbreviation if `translation_name == column_name`
  # (apply translation 'subject' to column 'subject' and save it to column 'subject_new')
  # (apply translation 'result' to column 'res' and save it to column 'res_new')
  df_new_overwritten <- lama_translate(
    df,
    dict,
    subject_new = subject,
    res_new = result(res)
  )
  str(df_new_overwritten)

#> 'data.frame':	5 obs. of  5 variables:
#>  $ pupil      : num  1 1 2 2 3
#>  $ subject    : Factor w/ 2 levels "en","ma": 1 2 2 1 1
#>  $ res        : num  1 2 3 2 2
#>  $ subject_new: Factor w/ 2 levels "English","Mathematics": 1 2 2 1 1
#>  $ res_new    : Factor w/ 3 levels "Very good","Good",..: 1 2 3 2 2

  # Example-4 (atomic vector): Translate a vector
  sub <- c("ma", "en", "ma")
  sub_new <- df_new_overwritten <- lama_translate(
    sub,
    dict,
    subject
  )
  str(sub_new)

#>  Factor w/ 2 levels "English","Mathematics": 2 1 2

  # Example-5 (factor): Translate a factor
  sub <- factor(c("ma", "en", "ma"), levels = c("ma", "en"))
  sub_new <- df_new_overwritten <- lama_translate(
    sub,
    dict,
    subject,
    keep_order = TRUE
  )
  str(sub_new)

#>  Factor w/ 2 levels "Mathematics",..: 1 2 1

  # Example-6 (data.frame): Using 'lama_translate_'
  # (apply translation 'subject' to column 'subject' and save it to column 'subject_new')
  # (apply translation 'result' to column 'res' and save it to column 'res_new')
  df_new <- lama_translate_(
    df,
    dict,
    translation = c("subject", "result"),
    col = c("subject", "res"),
    col_new = c("subject_new", "res_new")
  )
  str(df_new)

#> 'data.frame':	5 obs. of  5 variables:
#>  $ pupil      : num  1 1 2 2 3
#>  $ subject    : Factor w/ 2 levels "en","ma": 1 2 2 1 1
#>  $ res        : num  1 2 3 2 2
#>  $ subject_new: Factor w/ 2 levels "English","Mathematics": 1 2 2 1 1
#>  $ res_new    : Factor w/ 3 levels "Very good","Good",..: 1 2 3 2 2

  # Example-7 (atomic vector): Translate an integer vector
  res <- c(1, 2, 1, 3, 1, 2)
  res_new <- df_new_overwritten <- lama_translate_(
    res,
    dict,
    "result"
  )
  str(res_new)

#>  Factor w/ 3 levels "Very good","Good",..: 1 2 1 3 1 2

  # Example-8 (factor): Translate a factor
  sub <- factor(c("ma", "en", "ma"), levels = c("ma", "en"))
  sub_new <- df_new_overwritten <- lama_translate_(
    sub,
    dict,
    "subject",
    keep_order = TRUE
  )
  str(sub_new)

#>  Factor w/ 2 levels "Mathematics",..: 1 2 1